Initial support for 'imports' of other files with security annotations
The main problem to support 'imports' is to resolve security annotations
of referenced entities (eg. functions). After some iterations, we decide
to re-parse security annotations from the 'element' property.
It means that the role of the parser_visitor.dart is to detect security annotations
errors, but the computed security types (stored in the AstNode) is not used by
the security resolver.

In the security resolver we need to give a security type to an identifier. However
to access to the defining identifier's AstNode, we need to use element.computeNode().
The invocation of this function has two problems:
- If we execute the security analysis as a Task we get an error (because computeNode
triggers the execution of the task workflow again)
- If we do not use the Task model, the computeNode() method executes successfully, but
the result is a completely new AstNode (and hence it does not contain the security types
that the parser_visitor.dart parsed).

Conclusion: As a workaround, we re-parse the security types from the 'element' information
(assuming that there are no errors in the security annotations)